Split into multiple routes.
