+++ backend/auth/bearer.py
@@ -1,0 +1,10 @@
+from ninja.security import HttpBearer
+
+
+class AuthBearer(HttpBearer):
+    def authenticate(self, request, token):
+        # Auth stuff (e.g. JWT) can go here.
+        if token == "supersecret":
+            return "user123"
+        else:
+            return None
