+++ backend/backend/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from .api import api
+from endpoints.api import api
 
 urlpatterns = [
     path('admin/', admin.site.urls),
